app/schemas/document.py
from datetime import datetime

from pydantic import BaseModel, ConfigDict


class DocumentRead(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: int
    proposal_id: int
    display_name: str
    mime_type: str
    size_bytes: int
    created_at: datetime
