requirements.txt
alembic==1.16.5
bcrypt==4.3.0
email-validator==2.2.0
fastapi==0.116.1
python-multipart==0.0.20
pydantic-settings==2.11.0
psycopg[binary]==3.2.10
PyJWT==2.10.1
sqlalchemy==2.0.43
uvicorn[standard]==0.35.0
